+++ src/includes/classes/Utils/Shortcode.php
@@ -30,22 +30,166 @@ use function get_defined_vars as vars;
 class Shortcode extends SCoreClasses\SCore\Base\Core
 {
     /**
-     * Shortcode [if] name.
+     * `[if /]` shortcode name.
      *
-     * @since 160707.2545 Initial release.
+     * @since 16xxxx Refactor.
      *
-     * @param string Shortcode [if] name.
+     * @param string
      */
-    public $name;
+    public $name = '';
 
     /**
-     * Shortcode [else] name.
+     * `[else]` shortcode name.
      *
-     * @since 160707.2545 Initial release.
+     * @since 16xxxx Refactor.
      *
-     * @param string Shortcode [else] name.
+     * @param string
      */
-    public $else_name;
+    protected $else_name = '';
+
+    /**
+     * Initialized?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $initialized = false;
+
+    /**
+     * Multisite network?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $is_multisite = false;
+
+    /**
+     * WooCommerce installed?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $is_wc_installed = false;
+
+    /**
+     * Can `eval()`?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $can_eval = false;
+
+    /**
+     * `php=""` attribute enabled?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $enable_php_att = false;
+
+    /**
+     * `for_blog=""` attribute enabled?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $enable_for_blog_att = false;
+
+    /**
+     * Arbitrary attributes enabled?
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param bool
+     */
+    protected $enable_arbitrary_atts = false;
+
+    /**
+     * Whitelisted arbitrary attributes.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param array
+     */
+    protected $whitelisted_arbitrary_atts = [];
+
+    /**
+     * Debug att default.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param string
+     */
+    protected $debug_att_default = '';
+
+    /**
+     * Current user ID.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param int
+     */
+    protected $current_user_id = 0;
+
+    /**
+     * Current shortcode depth.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param int
+     */
+    protected $current_depth = 0;
+
+    /**
+     * Current shortcode.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param string
+     */
+    protected $current_shortcode = '';
+
+    /**
+     * Current raw attributes.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param array
+     */
+    protected $current_raw_atts = [];
+
+    /**
+     * Current attributes.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param array
+     */
+    protected $current_atts = [];
+
+    /**
+     * Current conditions.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param string
+     */
+    protected $current_conditions = '';
+
+    /**
+     * Current errors.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param array
+     */
+    protected $current_errors = [];
 
     /**
      * Class constructor.
@@ -63,57 +207,104 @@ class Shortcode extends SCoreClasses\SCore\Base\Core
     }
 
     /**
+     * Maybe initialize.
+     *
+     * @since 16xxxx Refactor.
+     */
+    protected function maybeInitialize()
+    {
+        if ($this->initialized) {
+            return; // Did this already.
+        }
+        $this->initialized = true;
+
+        $this->is_multisite    = is_multisite();
+        $this->is_wc_installed = defined('WC_VERSION');
+
+        $this->can_eval                   = c::canCallFunc('eval');
+        $this->enable_php_att             = (bool) s::getOption('enable_php_att');
+        $this->enable_for_blog_att        = (bool) s::getOption('enable_for_blog_att');
+        $this->enable_arbitrary_atts      = (bool) s::getOption('enable_arbitrary_atts');
+        $this->whitelisted_arbitrary_atts = $this->enable_arbitrary_atts ? preg_split('/\s+/u', s::getOption('whitelisted_arbitrary_atts'), -1, PREG_SPLIT_NO_EMPTY) : [];
+        $this->debug_att_default          = s::getOption('debug_att_default');
+    }
+
+    /**
      * `[if /]` shortcode.
      *
      * @since 160707.2545 Initial release.
      *
-     * @param array       $atts      Shortcode attributes.
-     * @param string|null $content   Shortcode content.
-     * @param string      $shortcode Shortcode name.
+     * @param array  $atts      Shortcode attributes.
+     * @param string $content   Shortcode content.
+     * @param string $shortcode Shortcode name.
      */
-    public function onShortcode(array $atts, $content, string $shortcode): string
+    public function onShortcode(array $atts, $content = '', $shortcode = ''): string
     {
-        $default_atts = [
-            'expr'                => '',
-            'is_user_logged_in'   => '',
-            'is_paying_customer'  => '',
-            'current_user_can'    => '', 'for_blog' => '0',
-            'current_user_option' => '', 'current_user_meta' => '',
-            'satisfy'             => 'all',
-        ];
-        $atts    = c::mbTrim($atts);
-        $atts    = shortcode_atts($default_atts, $atts, $shortcode);
-        $content = (string) $content; // Force string value.
-
-        $atts['expr'] = (string) $atts['expr'];
-        $atts['expr'] = $atts['expr'] ? c::unescHtml($atts['expr']) : '';
-
-        $atts['is_user_logged_in'] = (string) $atts['is_user_logged_in'];
-        $atts['is_user_logged_in'] = isset($atts['is_user_logged_in'][0]) ? filter_var($atts['is_user_logged_in'], FILTER_VALIDATE_BOOLEAN) : null;
-
-        $atts['is_paying_customer'] = (string) $atts['is_paying_customer'];
-        $atts['is_paying_customer'] = isset($atts['is_paying_customer'][0]) ? filter_var($atts['is_paying_customer'], FILTER_VALIDATE_BOOLEAN) : null;
-
-        $atts['current_user_can'] = (string) $atts['current_user_can'];
-        $atts['current_user_can'] = $atts['current_user_can'] ? c::unescHtml($atts['current_user_can']) : '';
-
-        $atts['current_user_option'] = (string) $atts['current_user_option'];
-        $atts['current_user_option'] = $atts['current_user_option'] ? c::unescHtml($atts['current_user_option']) : '';
-
-        $atts['current_user_meta'] = (string) $atts['current_user_meta'];
-        $atts['current_user_meta'] = $atts['current_user_meta'] ? c::unescHtml($atts['current_user_meta']) : '';
-
-        $atts['for_blog'] = (int) $atts['for_blog'];
-        $atts['satisfy']  = $atts['satisfy'] === 'any' ? 'any' : 'all';
-
-        $is_multisite    = is_multisite();
-        $current_user_id = (int) get_current_user_id();
-        $shortcode_depth = strspn($shortcode, '_'); // Based on a zero index.
-        $else_tag        = '['.str_repeat('_', $shortcode_depth).$this->else_name.']';
-        $conditions      = ''; // Initialize the full set of all conditions built below.
         /*
-         * Construct content for if/else conditions.
+         * Maybe initialize.
          */
+        $this->maybeInitialize();
+
+        /*
+         * Content/shortcode.
+         */
+        $content   = (string) $content;
+        $shortcode = (string) $shortcode;
+
+        /*
+         * Parse attributes.
+         */
+        $default_atts = [
+            // PHP attribute.
+            'php' => '', // PHP expression.
+
+            // WordPress user-specific attributes.
+            'current_user_is_logged_in' => '', // `true|false`.
+            'current_user_can'          => '', // Role/cap expression.
+            'current_user_option'       => '', // Meta key expression.
+            'current_user_meta'         => '', // Meta key expression.
+
+            // WooCommerce customer-specific attributes.
+            'current_user_is_paying_customer' => '', // `true|false`.
+            'current_user_bought_product'     => '', // Product ID (or SKU) expression.
+
+            // Attribute modifiers.
+            '_for_blog' => '0', // A specific blog ID.
+            '_satisfy'  => 'all', // `any` or `all` (default).
+
+            // Debugging attributes.
+            '_debug' => $this->debug_att_default,
+            // `1|on|yes|true|0|off|no|false` (or `verbose`).
+        ];
+        $raw_atts = $atts; // Copy.
+        $atts     = c::mbTrim(c::unescHtml($atts));
+        $atts     = array_merge($default_atts, $atts);
+
+        $atts['_for_blog'] = (int) $atts['_for_blog'];
+        $atts['_satisfy']  = $atts['_satisfy'] === 'any' ? 'any' : 'all';
+
+        if ($atts['_debug'] && $atts['_debug'] !== 'verbose') {
+            $atts['_debug'] = filter_var($atts['_debug'], FILTER_VALIDATE_BOOLEAN);
+        } elseif ($atts['_debug'] !== 'verbose') {
+            $atts['_debug'] = false;
+        }
+
+        /*
+         * Set 'current' properties.
+         */
+        $this->current_user_id    = (int) get_current_user_id();
+        $this->current_depth      = strspn($shortcode, '_');
+        $this->current_shortcode  = $shortcode;
+        $this->current_raw_atts   = $raw_atts;
+        $this->current_atts       = $atts;
+        $this->current_conditions = '';
+        $this->current_errors     = [];
+
+        /*
+         * Parse content into if/else conditions.
+         */
+        $else_tag = '['.str_repeat('_', $this->current_depth).$this->else_name.']';
+
         if (mb_strpos($content, $else_tag) !== false) {
             list($content_if, $content_else) = explode($else_tag, $content, 2);
             $content_if                      = c::htmlTrim($content_if);
@@ -122,212 +313,271 @@ class Shortcode extends SCoreClasses\SCore\Base\Core
             $content_if   = c::htmlTrim($content);
             $content_else = ''; // Default (empty).
         }
+
         /*
-         * Add conditions from the `expr=""` attribute, if applicable.
-         *
-         * - [if expr="current_user_can('access_pkg_slug1')" /]
-         * - [if expr="current_user_can('access_pkg_slug1') and current_user_can('access_pkg_slug2')" /]
-         * - [if expr="current_user_can('access_ccap_one') or get_user_option('meta_key') === 'value'" /]
+         * Initial validations.
          */
-        if ($atts['expr']) {
-            if (!s::getOption('enable_expr_att')) {
-                // This is disabled by default. If enabled, a filter can disable it on child sites of a network.
-                // add_filter('s2member_x_options', function(array $options) { $options['if_shortcode_expr_enable'] = '0'; return $options; });
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `expr=""` not enabled on this site.', 'if-shortcode'), $shortcode), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            }
-            if ($conditions) { // Taken as-is; raw PHP expression.
-                $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').'('.$atts['expr'].')';
-            } else {
-                $conditions = '('.$atts['expr'].')';
-            }
+        if (!$this->can_eval) {
+            $this->current_errors[] = sprintf(__('The `[%1$s]` shortcode requires the legitimate use of PHP `eval()`.', 'if-shortcode'), $this->current_shortcode).
+                ' '.__('Unfortunately, `eval()` is currently disabled by your PHP configuration. Therefore, the use of this shortcode will not be possible.', 'if-shortcode').
+                ' '.__('To resolve this issue, please contact your server administrator and ask them to enable `eval()` in PHP.', 'if-shortcode');
         }
-        /*
-         * Add conditions from the `is_user_logged_in=""` attribute, if applicable.
-         *
-         * - [if is_user_logged_in="true" /]
-         * - [if is_user_logged_in="yes" /]
-         * - [if is_user_logged_in="false" /], etc.
-         */
-        if (isset($atts['is_user_logged_in'])) {
-            if ($atts['is_user_logged_in'] === false) {
-                $_is_user_logged_in_condition = '!is_user_logged_in()';
-            } else { // Default behavior is a boolean true.
-                $_is_user_logged_in_condition = 'is_user_logged_in()';
-            }
-            if ($conditions) {
-                $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').$_is_user_logged_in_condition;
-            } else {
-                $conditions = $_is_user_logged_in_condition;
-            }
-            // unset($_is_user_logged_in_condition); // Housekeeping.
+        if ($this->current_atts['php'] && !$this->enable_php_att) {
+            $this->current_errors[] = sprintf(__('`[%1$s]` shortcode attribute `php="%2$s"` is not enabled via plugin options.', 'if-shortcode'), $this->current_shortcode, $this->current_atts['php']);
         }
-        /*
-         * Add conditions from the `is_paying_customer=""` attribute, if applicable.
-         *
-         * - [if is_paying_customer="true" /]
-         * - [if is_paying_customer="yes" /]
-         * - [if is_paying_customer="false" /], etc.
-         */
-        if (isset($atts['is_paying_customer'])) {
-            if ($atts['is_paying_customer'] === false) {
-                $_is_paying_customer_condition = '(!'.$current_user_id.' || !defined(\'WC_VERSION\') || !get_user_meta('.$current_user_id.', \'paying_customer\', true))';
-            } else { // Default behavior is a boolean true.
-                $_is_paying_customer_condition = '('.$current_user_id.' && defined(\'WC_VERSION\') && get_user_meta('.$current_user_id.', \'paying_customer\', true))';
-            }
-            if ($conditions) {
-                $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').$_is_paying_customer_condition;
-            } else {
-                $conditions = $_is_paying_customer_condition;
-            }
-            // unset($_is_paying_customer_condition); // Housekeeping.
+        if ($this->current_atts['_for_blog'] && !$this->enable_for_blog_att) {
+            $this->current_errors[] = sprintf(__('`[%1$s]` shortcode attribute `_for_blog="%2$s"` is not enabled via plugin options.', 'if-shortcode'), $this->current_shortcode, $this->current_atts['_for_blog']);
         }
-        /*
-         * Add conditions from the `current_user_can=""` attribute, if applicable.
-         *
-         * - [if current_user_can="access_pkg_slug1" /]
-         * - [if current_user_can="access_pkg_slug1 AND access_pkg_slug2" /]
-         * - [if current_user_can="(access_pkg_slug1 AND access_pkg_slug2) OR (access_ccap_one AND access_ccap_two)" /]
-         */
-        if ($atts['current_user_can']) {
-            if (mb_strpos($atts['current_user_can'], "'") !== false) {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_can="%2$s"` contains apostrophe.', 'if-shortcode'), $shortcode, $atts['current_user_can']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            } elseif (!preg_match('/\((?:(?:[^()]+)|(?R))*\)/u', '('.$atts['current_user_can'].')', $_m) || $_m[0] !== '('.$atts['current_user_can'].')') {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_can="%2$s"` contains unbalanced `()` brackets.', 'if-shortcode'), $shortcode, $atts['current_user_can']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            } elseif ($atts['for_blog'] && $is_multisite && !s::getOption('enable_for_blog_att')) {
-                // This is disabled by default. If enabled, a filter can disable it on child sites of a network.
-                // add_filter('woocommerce_s2member_x_options', function(array $options) { $options['if_shortcode_for_blog_enable'] = '0'; return $options; });
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `for_blog=""` not enabled on this site.', 'if-shortcode'), $shortcode), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            }
-            $_current_user_can_conditions = ''; // Initialize conditions.
-            $_previous_frag               = ''; // Initialize previous fragment.
-
-            foreach (preg_split('/([()])|\s+/u', $atts['current_user_can'], -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY) as $_frag) {
-                $_space = !$_current_user_can_conditions || !$_previous_frag || $_previous_frag === '(' || $_frag === '(' || $_frag === ')' ? '' : ' ';
-
-                if (in_array($_lc_frag = mb_strtolower($_frag), ['(', ')', 'and', 'or', '&&', '||'], true)) {
-                    $_frag = $_lc_frag === 'and' ? '&&' : $_frag;
-                    $_frag = $_lc_frag === 'or' ? '||' : $_frag;
-                    $_current_user_can_conditions .= $_space.$_frag;
-                } elseif ($atts['for_blog'] && $is_multisite) { // Network-only.
-                    $_current_user_can_conditions .= $_space.'current_user_can_for_blog('.$atts['for_blog'].', \''.$_frag.'\')';
-                } else { // Build the `current_user_can('[cap]')` check.
-                    $_current_user_can_conditions .= $_space.'current_user_can(\''.$_frag.'\')';
-                }
-                $_previous_frag = $_frag; // Previous fragment.
-            } // unset($_frag, $_lc_frag, $_previous_frag, $_space); // Housekeeping.
-
-            if ($_current_user_can_conditions) {
-                if ($conditions) {
-                    $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').'('.$_current_user_can_conditions.')';
-                } else {
-                    $conditions = '('.$_current_user_can_conditions.')';
-                }
-            } // unset($_current_user_can_conditions); // Housekeeping.
+        if (isset($this->current_atts[0])) { // i.e., Attributes that do not have an `=""` value are numerically indexed by WP.
+            $this->current_errors[] = sprintf(__('`[%1$s]` shortcode attribute names, by themselves, without `=""`, are not supported at this time; because that particular format depends on functionality which may change in a future version of WP.', 'if-shortcode'), $this->current_shortcode);
         }
+
         /*
-         * Add conditions from the `current_user_option=""` attribute, if applicable.
-         *
-         * - [if current_user_option="key" /]
-         * - [if current_user_option="key = value OR key = value" /]
-         * - [if current_user_option="(key != value AND key != value) OR (key AND another_key > 4)" /]
+         * Iterate all shortcode attributes.
          */
-        if ($atts['current_user_option']) {
-            if (mb_strpos($atts['current_user_option'], "'") !== false) {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_option="%2$s"` contains apostrophe.', 'if-shortcode'), $shortcode, $atts['current_user_option']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            } elseif (!preg_match('/\((?:(?:[^()]+)|(?R))*\)/u', '('.$atts['current_user_option'].')', $_m) || $_m[0] !== '('.$atts['current_user_option'].')') {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_option="%2$s"` contains unbalanced `()` brackets.', 'if-shortcode'), $shortcode, $atts['current_user_option']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            }
-            $_current_user_option_conditions = ''; // Initialize conditions.
-            $_previous_frag                  = ''; // Initialize previous fragment.
+        foreach (array_keys($this->current_atts) as $_att_key) {
+            switch ((string) $_att_key) {
+                /*
+                 * PHP attribute.
+                 */
 
-            foreach (preg_split('/([()])|\s+/u', $atts['current_user_option'], -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY) as $_frag) {
-                $_space = !$_current_user_option_conditions || !$_previous_frag || $_previous_frag === '(' || $_frag === '(' || $_frag === ')' ? '' : ' ';
-
-                if (in_array($_lc_frag = mb_strtolower($_frag), ['(', ')', 'and', 'or', '&&', '||'], true)) {
-                    $_frag = $_lc_frag === 'and' ? '&&' : $_frag;
-                    $_frag = $_lc_frag === 'or' ? '||' : $_frag;
-                    $_current_user_option_conditions .= $_space.$_frag;
-                } elseif (($_frag_parts = preg_split('/([<>!=]+)|\s+/u', $_frag, -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY))) {
-                    $_operator = $_frag_parts[1] ?? ''; // Defaults to a boolean condition below; i.e., when empty.
-                    $_operator = $_operator && preg_match('/^\=+$/u', $_operator) ? '===' : $_operator;
-                    $_operator = $_operator && preg_match('/^\!\=+$/u', $_operator) ? '!==' : $_operator;
-                    $_value    = $_operator && isset($_frag_parts[2]) ? $_frag_parts[2] : '';
-
-                    if (!in_array($_operator, ['', '==', '!=', '===', '!==', '<=', '>=', '<>', '>', '<'], true)) {
-                        trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_option` contains invalid operator: %2$s.', 'if-shortcode'), $shortcode, $_operator), E_USER_ERROR);
-                        return ''; // Return empty string in case of error handlers allowing this to slide.
+                /*
+                 * `php="[expr]"`
+                 */
+                case 'php':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->current_atts[$_att_key]);
                     }
-                    $_current_user_option_conditions .= $_space.($_operator ? '(string)' : '').'get_user_option(\''.$_frag_parts[0].'\', '.$current_user_id.')'.($_operator ? ' '.$_operator.' \''.$_value.'\'' : '');
-                }
-                $_previous_frag = $_frag; // Previous fragment.
-            } // unset($_frag, $_lc_frag, $_previous_frag, $_space, $_frag_parts, $_operator, $_value); // Housekeeping.
+                    break;
 
-            if ($_current_user_option_conditions) {
-                if ($conditions) {
-                    $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').'('.$_current_user_can_conditions.')';
-                } else {
-                    $conditions = '('.$_current_user_option_conditions.')';
-                }
-            } // unset($_current_user_option_conditions); // Housekeeping.
-        }
-        /*
-         * Add conditions from the `current_user_meta=""` attribute, if applicable.
-         *
-         * - [if current_user_meta="key" /]
-         * - [if current_user_meta="key = value OR key = value" /]
-         * - [if current_user_meta="(key != value AND key != value) OR (key AND another_key > 4)" /]
-         */
-        if ($atts['current_user_meta']) {
-            if (mb_strpos($atts['current_user_meta'], "'") !== false) {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_meta="%2$s"` contains apostrophe.', 'if-shortcode'), $shortcode, $atts['current_user_meta']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            } elseif (!preg_match('/\((?:(?:[^()]+)|(?R))*\)/u', '('.$atts['current_user_meta'].')', $_m) || $_m[0] !== '('.$atts['current_user_meta'].')') {
-                trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_meta="%2$s"` contains unbalanced `()` brackets.', 'if-shortcode'), $shortcode, $atts['current_user_meta']), E_USER_ERROR);
-                return ''; // Return empty string in case of error handlers allowing this to slide.
-            }
-            $_current_user_meta_conditions = ''; // Initialize conditions.
-            $_previous_frag                = ''; // Initialize previous fragment.
+                /*
+                 * WordPress user-specific attributes.
+                 */
 
-            foreach (preg_split('/([()])|\s+/u', $atts['current_user_meta'], -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY) as $_frag) {
-                $_space = !$_current_user_meta_conditions || !$_previous_frag || $_previous_frag === '(' || $_frag === '(' || $_frag === ')' ? '' : ' ';
-
-                if (in_array($_lc_frag = mb_strtolower($_frag), ['(', ')', 'and', 'or', '&&', '||'], true)) {
-                    $_frag = $_lc_frag === 'and' ? '&&' : $_frag;
-                    $_frag = $_lc_frag === 'or' ? '||' : $_frag;
-                    $_current_user_meta_conditions .= $_space.$_frag;
-                } elseif (($_frag_parts = preg_split('/([<>!=]+)|\s+/u', $_frag, -1, PREG_SPLIT_DELIM_CAPTURE | PREG_SPLIT_NO_EMPTY))) {
-                    $_operator = $_frag_parts[1] ?? ''; // Defaults to a boolean condition below; i.e., when empty.
-                    $_operator = $_operator && preg_match('/^\=+$/u', $_operator) ? '===' : $_operator;
-                    $_operator = $_operator && preg_match('/^\!\=+$/u', $_operator) ? '!==' : $_operator;
-                    $_value    = $_operator && isset($_frag_parts[2]) ? $_frag_parts[2] : '';
-
-                    if (!in_array($_operator, ['', '==', '!=', '===', '!==', '<=', '>=', '<>', '>', '<'], true)) {
-                        trigger_error(sprintf(__('`[%1$s /]` shortcode attribute `current_user_meta` contains invalid operator: %2$s.', 'if-shortcode'), $shortcode, $_operator), E_USER_ERROR);
-                        return ''; // Return empty string in case of error handlers allowing this to slide.
+                /*
+                 * `current_user_is_logged_in="[true|false]"`
+                 */
+                case 'current_user_is_logged_in':
+                    if ($this->current_atts[$_att_key]) {
+                        $_negating = $this->current_atts[$_att_key] === 'false' ? '!' : '';
+                        $this->appendConditions($_negating.'is_user_logged_in()');
                     }
-                    $_current_user_meta_conditions .= $_space.($_operator ? '(string)' : '').'get_user_meta('.$current_user_id.', \''.$_frag_parts[0].'\', true)'.($_operator ? ' '.$_operator.' \''.$_value.'\'' : '');
-                }
-                $_previous_frag = $_frag; // Previous fragment.
-            } // unset($_frag, $_lc_frag, $_previous_frag, $_space, $_frag_parts, $_operator, $_value); // Housekeeping.
+                    break;
 
-            if ($_current_user_meta_conditions) {
-                if ($conditions) {
-                    $conditions .= ($atts['satisfy'] === 'any' ? ' || ' : ' && ').'('.$_current_user_can_conditions.')';
-                } else {
-                    $conditions = '('.$_current_user_meta_conditions.')';
-                }
-            } // unset($_current_user_meta_conditions); // Housekeeping.
+                /*
+                 * `current_user_can="[expr]"`
+                 */
+                case 'current_user_can':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->simpleExpr($_att_key, function ($cap) {
+                            return $this->current_atts['_for_blog'] && $this->is_multisite
+                                ? 'current_user_can_for_blog('.$this->current_atts['_for_blog'].', '.c::sQuote($cap).')'
+                                : 'current_user_can('.c::sQuote($cap).')';
+                        }));
+                    }
+                    break;
+
+                /*
+                 * `current_user_option="[expr]"`
+                 */
+                case 'current_user_option':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->simpleExpr($_att_key, function ($option_key) {
+                            return '('.$this->current_user_id.' ? get_user_option('.c::sQuote($option_key).') : false)';
+                        }));
+                    }
+                    break;
+
+                /*
+                 * `current_user_meta="[expr]"`
+                 */
+                case 'current_user_meta':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->simpleExpr($_att_key, function ($meta_key) {
+                            return '('.$this->current_user_id.' ? get_user_meta('.$this->current_user_id.', '.c::sQuote($meta_key).', true) : false)';
+                        }));
+                    }
+                    break;
+
+                /*
+                 * WooCommerce customer-specific attributes.
+                 */
+
+                /*
+                 * `current_user_is_paying_customer="[true|false]"`
+                 */
+                case 'current_user_is_paying_customer':
+                    if ($this->current_atts[$_att_key]) {
+                        $_negating = $this->current_atts[$_att_key] === 'false' ? '!' : '';
+                        $this->appendConditions($_negating.'('.(int) $this->is_wc_installed.' && '.$this->current_user_id.' ? (bool) get_user_meta('.$this->current_user_id.', \'paying_customer\', true) : false)');
+                    }
+                    break;
+
+                /*
+                 * `current_user_bought_product="[expr]"`
+                 */
+                case 'current_user_bought_product':
+                    if ($this->current_atts[$_att_key]) {
+                        $this->appendConditions($this->simpleExpr($_att_key, function ($sku) {
+                            return '('.(int) $this->is_wc_installed.' && '.$this->current_user_id.' ? (bool) wc_customer_bought_product(\'\', '.$this->current_user_id.', '.a::class.'::wcProductIdBySku('.c::sQuote($sku).')) : false)';
+                        }));
+                    }
+                    break;
+
+                /*
+                 * Arbitrary custom attributes.
+                 */
+
+                /*
+                 * `[arbitrary]="[true|false]"` or `[arbitrary]="[expr]"`
+                 */
+                default: // NOTE: Arbitrary attributes must match `^[a-z][a-z0-9_]+[a-z0-9]$` to avoid compatibility issues in the future.
+                    // For instance, WordPress v4.5.3 allows an attribute to start with `!`, but if that changes in the future it could lead to
+                    // some conditionals breaking and/or not behaving as originally intended. Best to avoid things like this altogether.
+
+                    if (is_int($_att_key)) { // Do not allow integer attribute keys; i.e., those without an `=""` value.
+                        $this->current_errors[] = sprintf(__('`[%1$s]` shortcode attribute `%2$s`, by itself, without `=""`, is not supported. Instead, use `%2$s="true"`. Consult the shortcode documentation for assistance.', 'if-shortcode'), $this->current_shortcode, $this->current_atts[$_att_key]);
+                        //
+                    } elseif ($this->current_atts[$_att_key] && mb_strpos($_att_key, '_') !== 0) { // Skip over `_` modidifers.
+                        //
+                        if (!$this->enable_arbitrary_atts) {
+                            $this->current_errors[] = sprintf(__('Arbitrary `[%1$s]` shortcode attribute `%2$s="%3$s"` is not enabled via plugin options.', 'if-shortcode'), $this->current_shortcode, $_att_key, $this->current_atts[$_att_key]);
+                        } elseif ($this->whitelisted_arbitrary_atts && !in_array($_att_key, $this->whitelisted_arbitrary_atts, true)) {
+                            $this->current_errors[] = sprintf(__('Arbitrary `[%1$s]` shortcode attribute `%2$s="%3$s"` is not whitelisted via plugin options.', 'if-shortcode'), $this->current_shortcode, $_att_key, $this->current_atts[$_att_key]);
+                        } elseif (!preg_match('/^[a-z][a-z0-9_]+[a-z0-9]$/u', $_att_key)) { // Make this extra strict to avoid future compatibility issues.
+                            $this->current_errors[] = sprintf(__('Arbitrary `[%1$s]` shortcode attribute `%2$s="%3$s"` contains invalid chars in the attribute name. Must match: `^[a-z][a-z0-9_]+[a-z0-9]$`.', 'if-shortcode'), $this->current_shortcode, $_att_key, $this->current_atts[$_att_key]);
+                        //
+                        } elseif ($this->current_atts[$_att_key] === 'true' || $this->current_atts[$_att_key] === 'false') {
+                            $_negating                     = $this->current_atts[$_att_key] === 'false' ? '!' : '';
+                            $_arbitrary_att_bool_condition = s::applyFilters('arbitrary_att_bool_condition', $_negating.$_att_key.'()', $_att_key, $this->current_atts[$_att_key]);
+                            $this->appendConditions($_arbitrary_att_bool_condition); // Allows for a custom condition filter.
+                        } else { // Treat it as a simple expression.
+                            $_arbitrary_att_expr_callback = s::applyFilters('arbitrary_att_expr_callback', function ($args) use ($_att_key) {
+                                return $_att_key.'('.implode(', ', c::sQuote(explode(',', $args))).')'; // Default callback.
+                            }, $_att_key, $this->current_atts[$_att_key]); // Allows for a custom callback filter.
+
+                            $this->appendConditions($this->simpleExpr($_att_key, $_arbitrary_att_expr_callback));
+                        }
+                    }
+                    break;
+            }
+        } // unset($_att_key, $_negating, $_arbitrary_att_bool_condition, $_arbitrary_att_expr_callback); // Housekeeping.
+
+        /*
+         * Evaluate, if possible.
+         */
+
+        try { // We can catch problems in PHP 7+ via exception.
+            if (!$this->current_errors && $this->current_conditions && $this->can_eval) {
+                $conditions_true = c::phpEval('return ('.$this->current_conditions.');');
+            } else {
+                $conditions_true = false; // Force false; not possible.
+            }
+        } catch (\Throwable $eval_Exception) {
+            $conditions_true        = false; // Force false.
+            $this->current_errors[] = $eval_Exception->getMessage();
+        }
+
+        /*
+         * Verbose debug output (if enabled).
+         */
+        if ($this->current_atts['_debug'] === 'verbose') {
+            $debug_verbose = '<pre style="padding:1em; border-radius:.25em; max-width:100%; overflow:auto; max-width:100%; overflow:auto;">'.
+                                sprintf(__('<code>[%1$s]</code> Shortcode Verbose', 'if-shortcode'), esc_html($this->name)).'<br />'.esc_html($this->recreate()).'<br /><br />'.
+                                sprintf(__('<code>[%1$s]</code> Shortcode Atts', 'if-shortcode'), esc_html($this->name)).'<br />'.c::mbTrim(c::dump($this->current_atts, true), "\n").'<br /><br />'.
+                                sprintf(__('<code>[%1$s]</code> Shortcode Conditions', 'if-shortcode'), esc_html($this->name)).'<br />'.esc_html($this->current_conditions).
+                             '</pre>';
+        } else {
+            $debug_verbose = ''; // Default behavior (empty string).
+        }
+
+        /*
+         * Deal with errors (if applicable).
+         */
+        if ($this->current_errors) {
+            if ($this->current_atts['_debug']) {
+                $error_colors   = 'background:#b30000; color:#FFFFFF;';
+                $error_li_first = '<li style="background:inherit; color:inherit; margin:0; padding:0;">';
+                $error_li       = '<li style="background:inherit; color:inherit; margin:1em 0 0 0; padding:0;">';
+
+                return $debug_verbose.// Verbose debug output (if enabled).
+                        '<div style="'.esc_attr($error_colors).' padding:1em; border-radius:.25em;">'.
+                            '<h4 style="background:inherit; color:inherit; margin:0 0 .5em 0; padding:0 0 .5em 0; line-height:1em; border-bottom:1px solid;">'.
+                                sprintf(_n('<code>[%1$s]</code> Shortcode Error', '<code>[%1$s]</code> Shortcode Errors', count($this->current_errors), 'if-shortcode'), esc_html($this->name)).
+                            '</h4>'.
+                            '<pre style="padding:1em; margin:0 0 .5em 0; border-radius:.25em; max-width:100%; overflow:auto;">'.
+                                esc_html($this->recreate()).
+                            '</pre>'.
+                            '<ul style="margin:0 0 0 2em; padding:0; background:inherit; color:inherit;">'.
+                                $error_li_first.implode('</li>'.$error_li, c::markdown($this->current_errors, ['no_p' => true])).'</li>'.
+                            '</ul>'.
+                        '</div>';
+            } elseif ($this->App->Config->©debug['©enable']) {
+                debug(0, c::issue(vars(), implode("\n", $this->current_errors)));
+                trigger_error(implode("\n", $this->current_errors), E_USER_WARNING);
+            }
+            return ''; // Default behavior; fail silently with conditions forced to a false state.
         }
         /*
-         * Test the expression return value and deal with nested shortcodes.
-         * This uses all possible conditions from above and tests them in one `eval()`.
+         * Return shortcode output (if no errors above).
          */
-        return do_shortcode($conditions && c::phpEval('return ('.$conditions.');') ? $content_if : $content_else);
+        return $debug_verbose.// Verbose debug output (if enabled).
+               do_shortcode($conditions_true ? $content_if : $content_else);
+    }
+
+    /**
+     * Build conditions from simple expression.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param string   $att      Shortcode attribute name.
+     * @param callable $callback A callback handler.
+     */
+    protected function simpleExpr(string $att, callable $callback): string
+    {
+        if (!($conditions = c::simplePhpExpr($this->current_atts[$att] ?? '', $callback))) {
+            $this->current_errors[] = sprintf(__('`[%1$s]` shortcode attribute contains an invalid (or imbalanced) expression: `%2$s="%3$s"`', 'if-shortcode'), $this->current_shortcode, $att, $this->current_atts[$att]);
+        }
+        return $conditions;
+    }
+
+    /**
+     * Append conditions.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @param string $conditions Conditions.
+     */
+    protected function appendConditions(string $conditions)
+    {
+        if (!$conditions) {
+            return; // Nothing to do here.
+        } elseif ($this->current_conditions) { // Conditions already exist?
+            $this->current_conditions .= ($this->current_atts['_satisfy'] === 'any' ? ' || ' : ' && ').'('.$conditions.')';
+        } else {
+            $this->current_conditions = '('.$conditions.')';
+        }
+    }
+
+    /**
+     * Recreate shortcode.
+     *
+     * @since 16xxxx Refactor.
+     *
+     * @return string Shortcode.
+     */
+    protected function recreate(): string
+    {
+        $shortcode = '['.$this->current_shortcode;
+
+        foreach ($this->current_raw_atts as $_att_key => $_att_value) {
+            if (is_int($_att_key)) {
+                $shortcode .= ' '.$_att_value;
+            } else {
+                $shortcode .= ' '.$_att_key.'="'.$_att_value.'"';
+            }
+        } // unset($_att_key, $_att_value); // Housekeeping.
+
+        return $shortcode .= ']';
     }
 }