+++ src/includes/wp-sharks-core-rv.php
@@ -6,6 +6,6 @@ if (!defined('WPINC')) {
     exit('Do NOT access this file directly: '.basename(__FILE__));
 }
 $GLOBALS['wp_sharks_core_rv'] = [
-    'min' => '160706.64464', //wp-sharks-core-required-version//
+    'min' => '160709.27823', //wp-sharks-core-required-version//
     'max' => '', //wp-sharks-core-max-compatible-version//
 ];