+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## $v
+
+- Updating to latest WPSC and websharks/core.
+- Refactor. Now using Simple Expression syntax from websharks/core.
+- Adding support for a new `_debug=""` attribute.
+- Adding support for arbitrary attributes.
+
 ## v160707.2545
 
 - Updating to latest phings build system.